% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw3dMap.R
\name{drawColorTracks}
\alias{drawColorTracks}
\title{set up call to draw3dMap with colors for the tracks}
\usage{
drawColorTracks(paths, drawPalette = "default", colorMode = "track", ...)
}
\arguments{
\item{paths}{dataframe see draw3dMap}

\item{drawPalette}{name of the palette to be used.  Options are "plasma",
"magma","viridis","heat","rainbow","red-blue-green","red-blue","default"}

\item{colorMode}{character either "track" where each segment in the path has 
a different color sequentially chosen from the palette, or "speed" where the 
pieces of the path are colored based on the speed recorded.}

\item{...}{parameters passed to draw3dMap}
}
\description{
\code{drawColorTracks} draw 3D map from elevation and feature data,
   and add a set of colored paths to the plot
}
\details{
This is a wrapper for draw3dMap.
}
