% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featuresForElevations.R
\name{featuresForElevations}
\alias{featuresForElevations}
\title{create and save the layers of a rasterStack containing the feature data for 
  an elevation raster}
\usage{
featuresForElevations(rasterFileSetName, rasterDir, shapefileDir = NULL,
  USStatevec = NULL, CAProvincevec = NULL,
  featureDataSource = "Shapefiles", writeShapefiles = TRUE,
  includeAllRoads = FALSE, year = 2017, zeroBufferTowns = FALSE,
  zeroBufferWater = FALSE,
  workProj4 = "+proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0",
  mapbuffer = 0, mapmergebuffer = 0, sliceFeatureBuffer = 0,
  maxRasterize = 1e+05, polyClean = FALSE, polySimplify = 0,
  polyMethod = "vis", polyWeighting = 0.85, polySnapInt = 1e-04,
  silent = FALSE, noisy = FALSE, ...)
}
\arguments{
\item{rasterFileSetName}{vector of names of saved raster data files}

\item{rasterDir}{location to load and save raster files}

\item{shapefileDir}{location to load/save shapefiles}

\item{USStatevec}{vector of standard 2-letter abbreviations of US states, or
pre-defined regional aggregations of states for use in defining map or
finding features to include in the map}

\item{CAProvincevec}{vector of standard 2-letter abbreviations of Canadian
provinvces states, or pre-defined regional aggregations of states for 
use in defining map or finding features to include in the map}

\item{featureDataSource}{"Shapefiles" to load saved shapefiles, "TIGER"
to fetch TIGER data for US states, "Raster" to load saved raster data
from directory specified}

\item{writeShapefiles}{write/overwrite shapefiles if TIGER data is used}

\item{includeAllRoads}{include all roads in shapefile, not just highways}

\item{year}{numeric year to use in calls for map boundaries/features}

\item{zeroBufferTowns}{use zero buffer trick to repair town polygon shapefile}

\item{zeroBufferWater}{use zero buffer trick to repair water polygon shapefile}

\item{workProj4}{coordinte reference projection string}

\item{mapbuffer}{numeric value to expand defined map}

\item{mapmergebuffer}{numeric value to expand components of the 
map defined before merging, to eliminate gaps from minor
boundary inconsistencies}

\item{sliceFeatureBuffer}{numeric value to expand the extent of the area
used to estrict features befire rasterization}

\item{maxRasterize}{number of items for calls to velox$rasterize}

\item{polyClean}{fix topology errors}

\item{polySimplify}{amount of polygon simplification, see help
for rmapshaper::ms_simplify}

\item{polyMethod}{simplification method either "vis" or "dp"}

\item{polyWeighting}{see help for rmapshaper::ms_simplify}

\item{polySnapInt}{see help for rmapshaper::ms_simplify}

\item{silent}{logical, suppress most output}

\item{noisy}{logical, more output to track progress}

\item{...}{values passed through}
}
\description{
\code{featuresForElevations} create and save the laters of a rasterStack 
  containing the feature data for an elevation raster
}
