% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featuresForElevations.R
\name{imageForRasters}
\alias{imageForRasters}
\title{create and save a rasterLayer containing the image data fetched by 
  openStreetmaps for an elevation raster}
\usage{
imageForRasters(rasterFileSetName, rasterDir, imageSource)
}
\arguments{
\item{rasterFileSetName}{vector of names of saved raster data files}

\item{rasterDir}{location to load and save raster files}

\item{imageSource}{string containing name for OpenStreetMap map type
from c("bing","maptoolkit-topo","nps","apple-iphoto")}
}
\description{
\code{ImageForElevations} create and save the laters of a rasterStack 
  containing the feature data for an elevation raster
}
