% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maptrack3d.R
\name{maptrack3d}
\alias{maptrack3d}
\title{maptrack3d: A package for drawing maps as 3d surfaces and displaying
  GPS track data on the surfaces.}
\description{
maptrack3d: A package for drawing maps as 3d surfaces and displaying
  GPS track data on the surfaces.
}
\section{main entry call}{

  \link{draw3dMap}
}

\section{read SRTM data saving rasters}{

  \link{elevationsToRaster}
}

\section{read shapefiles and create feature rasters form elevation rasters}{

  \link{featuresForElevations}
}

